Make create_script return None if script set up to immediately fire/stop. Resolves #1663.
